fake_job_detector/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
scikit-learn==1.3.0
python-whois==0.8.0
reportlab==4.0.4
Werkzeug==2.3.0
numpy>=1.21.0
scipy>=1.7.0
joblib>=1.0.0
threading-timer>=0.1.0
